[INCOMPLETE] Remove unused PulpTaskStatus class
The Pulp 2 task status mechanism is removed and the parts that are
inherited by PulpSyncStatus are moved over.

This may need a DB migration to remove the model.
